made input device view default
